requirements.txt
streamlit
sqlalchemy
openai
python-dotenv
